LayoutTests/imported/w3c/web-platform-tests/css/css-counter-styles/counter-style-additive-symbols-syntax-expected.txt


FAIL @counter-style 'additive-symbols: 1 "X"' is valid assert_equals: expected (undefined) undefined but got (string) "1 \"X\""
FAIL @counter-style 'additive-symbols: "X" 1' is valid assert_equals: expected (undefined) undefined but got (string) "1 \"X\""
FAIL @counter-style 'additive-symbols: 5 "V", 1 "I"' is valid assert_equals: expected (undefined) undefined but got (string) "5 \"V\", 1 \"I\""
PASS @counter-style 'additive-symbols: ' is invalid
PASS @counter-style 'additive-symbols: -1 "X"' is invalid
PASS @counter-style 'additive-symbols: 1 "I", 5 "V"' is invalid
PASS @counter-style 'additive-symbols: 1 "X", 1 "Y"' is invalid
